requirements.txt
# Core dependencies for Unified Story Inpaint Pipeline
torch>=2.0.0
torchvision>=0.15.0
diffusers>=0.30.2
transformers>=4.35.0
accelerate>=0.20.0

# Image processing
Pillow>=9.0.0
numpy>=1.21.0

# Optional: For better performance
xformers>=0.0.20
flash-attn>=2.0.0

# Development dependencies (optional)
jupyter>=1.0.0
matplotlib>=3.5.0
tqdm>=4.64.0

# Note: You may need to install specific versions based on your CUDA version
# For CUDA 11.8: torch==2.0.1+cu118
# For CUDA 12.1: torch==2.1.0+cu121
